# docs/swagger/swagger.yaml
basePath: /
definitions:
  APIErrorInfo:
    properties:
      code:
        example: validation_error
        type: string
      details:
        additionalProperties: true
        type: object
      message:
        example: Validation failed
        type: string
    type: object
  APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/APIErrorInfo'
      success:
        example: true
        type: boolean
      timestamp:
        example: "2025-01-15T10:30:00Z"
        type: string
    type: object
...
  CreateSessionRequest:
    properties:
      name:
        example: my-session
        maxLength: 100
        minLength: 1
        type: string
      qrCode:
        example: true
        type: boolean
      settings:
        $ref: '#/definitions/SessionSettings'
    required:
    - name
    type: object
...
  ErrorResponse:
    properties:
      error:
        example: validation_error
        type: string
      message:
        example: name is required
        type: string
    type: object
...
  PairPhoneRequest:
    properties:
      phone:
        example: "5511999999999"
        type: string
    required:
    - phone
    type: object
  PairPhoneResponse:
    properties:
      linkingCode:
        example: ABCD-EFGH
        type: string
    type: object
  ProxySettings:
    properties:
      enabled:
        example: true
        type: boolean
      host:
        example: proxy.example.com
        type: string
      pass:
        example: proxyPass123
        type: string
      port:
        example: "8080"
        type: string
      type:
        enum:
        - http
        - https
        - socks5
        example: http
        type: string
      user:
        example: proxyUser123
        type: string
    type: object
  QRCodeResponse:
    properties:
      expiresAt:
        example: "2025-01-15T10:35:00Z"
        type: string
      qrCode:
        example: 2@abc123...
        type: string
      qrCodeBase64:
        example: data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAA...
        type: string
      status:
        example: generated
        type: string
    type: object
...
  SessionListInfo:
    properties:
      connected:
        example: true
        type: boolean
      connectedAt:
        example: "2025-01-15T10:32:00Z"
        type: string
      createdAt:
        example: "2025-01-15T10:30:00Z"
        type: string
      deviceJid:
        example: 5511999999999@s.whatsapp.net
        type: string
      lastSeen:
        example: "2025-01-15T10:35:00Z"
        type: string
      name:
        example: my-session
        type: string
      sessionId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      settings:
        $ref: '#/definitions/SessionSettings'
      status:
        example: connected
        type: string
      updatedAt:
        example: "2025-01-15T10:35:00Z"
        type: string
    type: object
  SessionResponse:
    properties:
      connected:
        example: true
        type: boolean
      connectedAt:
        example: "2025-01-15T10:32:00Z"
        type: string
      createdAt:
        example: "2025-01-15T10:30:00Z"
        type: string
      deviceJid:
        example: 5511999999999@s.whatsapp.net
        type: string
      lastSeen:
        example: "2025-01-15T10:35:00Z"
        type: string
      name:
        example: my-session
        type: string
      qrCode:
        example: 2@abc123...
        type: string
      qrCodeBase64:
        example: data:image/png;base64,iVBORw0KGgo...
        type: string
      qrCodeExpiresAt:
        example: "2025-01-15T10:35:00Z"
        type: string
      sessionId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      settings:
        $ref: '#/definitions/SessionSettings'
      status:
        example: connected
        type: string
      updatedAt:
        example: "2025-01-15T10:35:00Z"
        type: string
    type: object
  SessionSettings:
    properties:
      proxy:
        $ref: '#/definitions/ProxySettings'
      webhook:
        $ref: '#/definitions/WebhookSettings'
    type: object
  SessionStatusResponse:
    properties:
      connected:
        example: true
        type: boolean
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      message:
        example: Session is already connected
        type: string
      status:
        example: connected
        type: string
    type: object
...
  WebhookSettings:
    properties:
      enabled:
        example: true
        type: boolean
      events:
        example:
        - Message
        - Receipt
        - Connected
        items:
          type: string
        type: array
      secret:
        example: supersecrettoken123
        type: string
      url:
        example: https://api.example.com/webhook
        type: string
    type: object
...
info:
  contact:
    email: support@zpwoot.com
    name: zpwoot Support
    url: https://github.com/your-org/zpwoot
  description: |-
    A comprehensive WhatsApp Business API built with Go, following Clean Architecture principles.
    Provides endpoints for session management, messaging, contacts, groups, media handling, and integrations.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: zpwoot WhatsApp API
  version: 1.0.0
paths:
...
  /sessions:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all WhatsApp sessions with their current status
        (without QR codes)
      produces:
      - application/json
      responses:
        "200":
          description: List of sessions (without QR codes)
          schema:
            $ref: '#/definitions/APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List WhatsApp Sessions
      tags:
      - Sessions
    post:
      consumes:
      - application/json
      description: Creates a new WhatsApp session with the specified configuration
      parameters:
      - description: Session configuration
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/CreateSessionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Session created successfully
          schema:
            $ref: '#/definitions/SessionResponse'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Session already exists
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create WhatsApp Session
      tags:
      - Sessions
  /sessions/{sessionId}:
    delete:
      consumes:
      - application/json
      description: Permanently deletes a WhatsApp session and all associated data
      parameters:
      - description: Session ID
        format: uuid
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Session deleted successfully
          schema:
            $ref: '#/definitions/SessionResponse'
        "400":
          description: Invalid session ID
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Deletion error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete WhatsApp Session
      tags:
      - Sessions
    get:
      consumes:
      - application/json
      description: Retrieves detailed information about a specific WhatsApp session
        (without QR code)
      parameters:
      - description: Session ID
        format: uuid
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Session information (without QR code)
          schema:
            $ref: '#/definitions/SessionListInfo'
        "400":
          description: Invalid session ID
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get WhatsApp Session
      tags:
      - Sessions
  /sessions/{sessionId}/connect:
    post:
      consumes:
      - application/json
      description: Connects a WhatsApp session. If already connected, returns current
        status with appropriate message.
      parameters:
      - description: Session ID
        format: uuid
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Session connected successfully or already connected
          schema:
            $ref: '#/definitions/SessionStatusResponse'
        "400":
          description: Invalid session ID
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Connection error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Connect WhatsApp Session
      tags:
      - Sessions
...
  /sessions/{sessionId}/disconnect:
    post:
      consumes:
      - application/json
      description: Disconnects an active WhatsApp session temporarily. If already
        disconnected, returns current status with appropriate message.
      parameters:
      - description: Session ID
        format: uuid
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Session disconnected successfully or already disconnected
          schema:
            $ref: '#/definitions/SessionStatusResponse'
        "400":
          description: Invalid session ID
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Disconnection error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Disconnect WhatsApp Session
      tags:
      - Sessions
...
  /sessions/{sessionId}/logout:
    post:
      consumes:
      - application/json
      description: Logs out a WhatsApp session permanently. Unlinks device from WhatsApp.
        Requires QR scan to reconnect.
      parameters:
      - description: Session ID
        format: uuid
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Session logged out successfully
          schema:
            $ref: '#/definitions/SessionResponse'
        "400":
          description: Invalid session ID
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Logout error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Logout WhatsApp Session
      tags:
      - Sessions
...
  /sessions/{sessionId}/pair:
    post:
      consumes:
      - application/json
      description: Pair a phone number without QR code
      parameters:
      - description: Session ID
        in: path
        name: sessionId
        required: true
        type: string
      - description: Phone number
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/PairPhoneRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/PairPhoneResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Pair phone
      tags:
      - Sessions
...
  /sessions/{sessionId}/qr:
    get:
      consumes:
      - application/json
      description: Retrieves the QR code for WhatsApp session authentication. Scan
        with WhatsApp mobile app.
      parameters:
      - description: Session ID
        format: uuid
        in: path
        name: sessionId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: QR code data
          schema:
            $ref: '#/definitions/QRCodeResponse'
        "400":
          description: Invalid session ID
          schema:
            $ref: '#/definitions/ErrorResponse'
        "404":
          description: Session not found
          schema:
            $ref: '#/definitions/ErrorResponse'
        "409":
          description: Session already connected
          schema:
            $ref: '#/definitions/ErrorResponse'
        "500":
          description: QR code generation error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get QR Code
      tags:
      - Sessions
...
securityDefinitions:
  ApiKeyAuth:
    description: 'API key for authentication. Use the value from .env file (API_KEY
      variable). Send as ''Authorization: your-api-key'' (without Bearer prefix)'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...